DRY account and assets balance date ranges
