+++ sveltekit/src/routes/account/[slug]/+page.server.ts
@@ -1,11 +1,10 @@
 import prisma from '$lib/helpers/prisma.server';
 import { getSelectAccountTypes } from '$lib/helpers/forms.server';
 import { selectBalanceGroups } from '$lib/helpers/forms';
-import { SortOrder } from '$lib/helpers/constants';
 import { notFound } from '$lib/helpers/misc';
 import { handlePeriodEnd, setChartDatasetColor } from '$lib/helpers/charts';
 import { eachWeekOfInterval, startOfWeek } from 'date-fns';
-import { getAccountCurrentBalance } from '$lib/helpers/models.server';
+import { getAccountBalanceDateRange, getAccountCurrentBalance } from '$lib/helpers/models.server';
 import type { ChartDataset } from 'chart.js';
 import type { Account } from '@prisma/client';
 
@@ -30,26 +29,7 @@ export const load = async ({ params }: { params: Params }) => {
 	// Generate chart dataset
 	const labels: string[] = [];
 	const balanceHistoryDataset: ChartDataset = { label: account.name, data: [] };
-	const queryByAccount = { where: { accountId: account.id } };
-
-	let periodStart: Date | undefined;
-	let periodEnd: Date | undefined;
-
-	if (account.isAutoCalculated) {
-		// When the account is auto-calculated, the balance history is based on the earliest and most recent transaction date
-		const earliestQuery = { ...queryByAccount, orderBy: { date: SortOrder.ASC } };
-		const latestQuery = { ...queryByAccount, orderBy: { date: SortOrder.DESC } };
-
-		periodStart = await prisma.transaction.findFirst(earliestQuery).then((t) => t?.date);
-		periodEnd = await prisma.transaction.findFirst(latestQuery).then((t) => t?.date);
-	} else {
-		// When the account is NOT auto-calculated, the balance history is based on the earliest and most recent balance statement
-		const earliestQuery = { ...queryByAccount, orderBy: { createdAt: SortOrder.ASC } };
-		const latestQuery = { ...queryByAccount, orderBy: { createdAt: SortOrder.DESC } };
-
-		periodStart = await prisma.accountBalanceStatement.findFirst(earliestQuery).then((abs) => abs?.createdAt); // prettier-ignore
-		periodEnd = await prisma.accountBalanceStatement.findFirst(latestQuery).then((abs) => abs?.createdAt); // prettier-ignore
-	}
+	const { periodStart, periodEnd } = await getAccountBalanceDateRange(account);
 
 	if (periodStart && periodEnd) {
 		const weeksInPeriod = eachWeekOfInterval({
